+++ ErsatzTV.Core.Tests/Scheduling/ClassicScheduling/GetStartTimeAfterTests.cs
@@ -1,5 +1,6 @@
 using ErsatzTV.Core.Domain;
 using ErsatzTV.Core.Scheduling;
+using Microsoft.Extensions.Logging;
 using NUnit.Framework;
 using Shouldly;
 
@@ -26,7 +27,8 @@ public class GetStartTimeAfterTests
             0,
             DateTimeOffset.Parse("2025-11-02T00:00:00-05:00"));
 
-        DateTimeOffset result = PlayoutModeSchedulerBase<ProgramScheduleItem>.GetStartTimeAfter(state, scheduleItem);
+        DateTimeOffset result =
+            PlayoutModeSchedulerBase<ProgramScheduleItem>.GetStartTimeAfter(state, scheduleItem, Option<ILogger>.None);
 
         result.ShouldBe(DateTimeOffset.Parse("2025-11-02T02:00:00-06:00"));
     }
@@ -55,7 +57,8 @@ public class GetStartTimeAfterTests
             0,
             DateTimeOffset.Parse("2025-08-29T00:10:00-05:00"));
 
-        DateTimeOffset result = PlayoutModeSchedulerBase<ProgramScheduleItem>.GetStartTimeAfter(state, scheduleItem);
+        DateTimeOffset result =
+            PlayoutModeSchedulerBase<ProgramScheduleItem>.GetStartTimeAfter(state, scheduleItem, Option<ILogger>.None);
 
         result.ShouldBe(DateTimeOffset.Parse("2025-08-29T00:10:00-05:00"));
     }